Create sum.cpp
Your program should ask the user for a sequence of numbers until the user inputs 0 (zero) and then the program should print the sum of postive numbers that the user input. Here are some example dialogs:
  Input sequence of integers (zero to stop):-4 -3 -20 10 45 -100 1 0
  The sum of postive numbers is: 56
  Input sequence of integers (zero to stop):-1 -3 -5 -7 8 10 0
  The sum of postive numbers is: 18
  Input sequence of integers (zero to stop):-3 -2 -1 1 2 3 0
  The sum of postive numbers is: 6
You are not allowed to use functions, arrays or other constructs that we have not yet studied.