Create age.cpp
It is the year 1000. A handsome prince is locked up in an enchanted castle guarded by a fearsome dragon. Every five years, a fair maiden is allowed to come to the castle to try to free the prince. To free him, the maiden has to guess the dragon's age. If the maiden guesses correctly, the prince is set free, they get married and live happily ever after. If the maiden guesses incorrectly, the dragon says whether he is younger or older and the maiden has to come back in five years.