Create figures.cpp
This project shall contain multiple files. Write a program that repeatedly asks the user to select figure size, paint character and then the figure. The figures are a square, a slash or a X. For a slash, the program needs to clarify if it is a forward or back slash. For a square, it needs to clarify if it is a filled or hollow square. Then, the program should print the requested figure. The program should quit if the user inputs an invalid option.